refactor(scripts): decompose baseline-attribution into phases/
Extracts the four phases of the baseline-attribution pipeline plus
their private helpers from the monolithic baseline-attribution-wiring.js
(912 LOC) into a sibling phases/ directory under
.agents/scripts/lib/orchestration/story-close/baseline-attribution/:

  - refresh-commit.js          (324 LOC) — buildKindScorer,
                                            stageAndCheckBaselineDrift,
                                            runRefreshCommit
  - gate-failure.js            (297 LOC) — DEFAULT_GATE_REGISTRY,
                                            handleBaselineGateFailure,
                                            runPreMergeGatesWithAttribution
  - regression-projection.js   (266 LOC) — diffCrapBaselines,
                                            projectCrapRegressions,
                                            projectRegressionsForGate, PROJECTORS
  - scope-discovery.js         ( 50 LOC) — computeStoryDiffPaths,
                                            validateProjectionContext

The parent baseline-attribution-wiring.js shrinks to 79 LOC and becomes
a thin re-export module, preserving the existing import surface so the
6 consumers (5 tests + auto-refresh-runner.js + phases/gates.js) keep
importing from the same path unchanged.

Pure code-move refactor; no behavior change. 6751/6751 tests pass,
including all 104 targeted baseline-attribution tests.

Refs #3002 #2988
