+++ .agents/scripts/lib/orchestration/story-close/baseline-attribution/phases/scope-discovery.js
@@ -1,0 +1,50 @@
+/**
+ * scope-discovery.js — phase 1 of baseline-attribution.
+ *
+ * Resolve the Story branch's diff scope vs `origin/<epicBranch>` so the
+ * downstream projection + attribution layers can split "this Story's
+ * regressions" from "sibling drift inherited from the Epic branch".
+ *
+ * Pure helpers — every git invocation goes through the injected
+ * `gitRunner` so tests can pin the diff math without spawning git.
+ */
+
+import { gitSpawn as defaultGitSpawn } from '../../../../git-utils.js';
+
+/**
+ * Compute repo-relative paths the Story branch changed vs `origin/<epicBranch>`.
+ * Best-effort: a non-zero diff exit returns an empty array so the caller
+ * conservatively treats every regression as non-attributable (the safer
+ * default — the close blocks rather than absorbs sibling drift).
+ */
+export function computeStoryDiffPaths({
+  cwd,
+  epicBranch,
+  storyBranch,
+  gitRunner = { gitSpawn: defaultGitSpawn },
+}) {
+  if (!cwd || !epicBranch || !storyBranch) return [];
+  const diff = gitRunner.gitSpawn(
+    cwd,
+    'diff',
+    '--name-only',
+    `origin/${epicBranch}...${storyBranch}`,
+  );
+  if (diff.status !== 0) return [];
+  return (diff.stdout || '')
+    .split('\n')
+    .map((line) => line.trim())
+    .filter((line) => line.length > 0)
+    .map((line) => line.replace(/\\/g, '/'));
+}
+
+/**
+ * Guard for the projection phase — the projectors require a worktree + the
+ * pair of branch refs to compute a meaningful diff.
+ */
+export function validateProjectionContext({ cwd, epicBranch, storyBranch }) {
+  if (!cwd) return false;
+  if (!epicBranch) return false;
+  if (!storyBranch) return false;
+  return true;
+}
